:recycle: Fix unit tests and update url link for receive LSK from /wallet to /wallet/send
