+++ src/components/sendTo/sendTo.test.js
@@ -41,13 +41,13 @@ describe('SendTo Component', () => {
     });
 
     it('renders correct link', () => {
-      expect(wrapper.find('Link').prop('to')).to.equal(`${routes.wallet.path}?recipient=${props.address}`);
+      expect(wrapper.find('Link').prop('to')).to.equal(`${routes.send.path}?recipient=${props.address}`);
     });
 
     it('updates when address changes', () => {
       wrapper.setProps({ address: '9876L' });
       wrapper.update();
-      expect(wrapper.find('Link').prop('to')).to.equal(`${routes.wallet.path}?recipient=9876L`);
+      expect(wrapper.find('Link').prop('to')).to.equal(`${routes.send.path}?recipient=9876L`);
     });
 
     it('renders delegate username', () => {
